BidList Crud

+++ src/main/java/org/ocr/poseidon/services/AbstractCrudService.java
@@ -1,16 +1,18 @@
 package org.ocr.poseidon.services;
 
-import org.ocr.poseidon.domain.BidList;
-import org.ocr.poseidon.domain.CrudEntity;
+import org.ocr.poseidon.exceptions.ItemNotFoundException;
+import org.ocr.poseidon.interfaces.CrudEntity;
+import org.ocr.poseidon.interfaces.CrudService;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 public abstract class AbstractCrudService<E extends CrudEntity<E>> implements CrudService<E> {
 
     protected final JpaRepository<E, Integer> repository;
 
-    protected AbstractCrudService(JpaRepository<E, Integer> repository){
+    protected AbstractCrudService(JpaRepository<E, Integer> repository) {
         this.repository = repository;
     }
 
@@ -21,7 +23,8 @@ public abstract class AbstractCrudService<E extends CrudEntity<E>> implements Cr
 
     @Override
     public E getById(Integer id) {
-        return repository.findById(id).orElseThrow();
+        return repository.findById(id)
+                .orElseThrow(() -> new ItemNotFoundException("Item id n°" + id + "is not found"));
     }
 
     @Override
@@ -31,15 +34,22 @@ public abstract class AbstractCrudService<E extends CrudEntity<E>> implements Cr
 
     @Override
     public E save(E entity) {
-        if(repository.existsById(entity.getId())){
+        if (repository.existsById(entity.getId())) {
             throw new RuntimeException();
         }
         return repository.save(entity);
     }
 
 
+    //todo factoriser le findById
     @Override
     public void delete(Integer id) {
-        repository.deleteById(id);
+
+        Optional<E> resultOptional = repository.findById(id);
+        if (resultOptional.isPresent()) {
+            repository.deleteById(id);
+        } else {
+            throw new ItemNotFoundException("Item id n°" + id + "is not found");
+        }
     }
 }
